fix: refactor nidcpower session apis integration tests
