+++ tests/integration/session_management/__init__.py
@@ -1,0 +1,1 @@
+"""Integration tests for driver-specific session management APIs."""
